% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnFilter.R
\name{columnFilterUI}
\alias{columnFilterUI}
\alias{columnFilter}
\title{Shiny module om een kolom te filteren}
\usage{
columnFilterUI(id, data, preset = NULL)

columnFilter(input, output, session, data, preset = NULL)
}
\description{
Shiny module om een kolom te filteren
}
