% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot.R
\name{custom_plot}
\alias{custom_plot}
\title{A parameterized interface to ggplot2}
\usage{
custom_plot(plotarguments, data, interactive)
}
\arguments{
\item{plotarguments}{A list of plot settings. Inspect code to see full list.}

\item{interactive}{A list of settings from interactive elements in the plot widget.}
}
\value{
A ggplot2 object
}
\description{
This function is for internal use only and very limited in scope.
Makes a ggplot2 plot based on a list of inputs, which may be gathered from a shiny application.
}
