% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplotcontrols_server.R
\name{customplotcontrols}
\alias{customplotcontrols}
\title{Shinto dashboard maker, server function}
\usage{
customplotcontrols(input, output, session, data_key, datasets, args = NULL)
}
\description{
Shinto dashboard maker, server function
}
