% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplotcontrols_ui.R
\name{customplotcontrolsUI}
\alias{customplotcontrolsUI}
\title{Shinto dashboard maker, UI function}
\usage{
customplotcontrolsUI(id, args = NULL, data_key, datasets)
}
\description{
Shinto dashboard maker, UI function
}
