% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_widget.R
\name{insert_widget}
\alias{insert_widget}
\title{Inserts a widget}
\usage{
insert_widget(
  id,
  args,
  datasets,
  selector = "#placeholder",
  where = "beforeEnd",
  buttons = c("close", "edit"),
  size = list(width = 500, height = 450, margin = 10, padding = 25, padding_bottom =
    100)
)
}
\description{
Wrapper around insertUI(widgetUI(...)) and callModule(widget, ...).
}
