% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_widget.R
\name{insert_saved_widgets}
\alias{insert_saved_widgets}
\title{Loads all widgets from a JSON}
\usage{
insert_saved_widgets(
  dash,
  datasets,
  buttons = "",
  session = getDefaultReactiveDomain(),
  ...
)
}
\description{
Loads all widgets from a JSON
}
